requirements.txt
paho_mqtt
psycopg2
psycopg2-binary
python-dotenv
